Implemented the Peer requests and started the Client. (The Peer file sucks right now, there's TONS of useless copying, it's just ugly)

+++ Torrent.cpp
@@ -5,7 +5,7 @@ Torrent::Torrent(std::string torrentFileName) : torrentFileName(torrentFileName)
     parseTorrentFile();
 }
 
-float Torrent::size() {
+int Torrent::size() {
     return 0;
 }
